renamed pieces to piece
